Alpha release polish (#17)
* Fix delete all files functionality

* Fix timeline card date subheader

* Get rid of alert statements in DriveHelper

* Fix date increment bug when adding diary entry

* Update timeline after deleting entry via card

* Improved deletion process

* Pin to Password

* Remove unimplemented style options

* Tallies on Insights Page

* Blank weather

* Entry snackbar

* Mobile-friendly password

* Mobile-friendly bottom settings grid

* Address some ES Lint complaints

* Proper mobile sidebar font

* Better diary update

* Properly show and sort dates

* Entry Viewer refactor

* Finish diary entry view refactor

* Improved scrolling mechanics

* Update entry store upon sign in

* Update node modules

* Limit to body text preview

* Fixed title styling

* Use jpg logo for Edge

* Better sidebar

* Update .gitignore

* Year old diary entry insight

* Entry navigation buttons

* Replace mood with thumbs

* Disable fullscreen for adding views

* Timeline filters

* Better card titles

* Update manifest.json

* Update node modules

+++ src/components/helpers/driveHelper.js
@@ -179,21 +179,41 @@ export default class DriveHelper {
      * 
      */
     static deleteAllFiles() {
-        gapi.client.drive.files.list({
-            'spaces': 'appDataFolder',
-            'fields': "nextPageToken, files(id, name)",
-            'pageSize': 25
-        }).then((response) => {
-            const files = response.result.files;
-            console.log("Files: " + JSON.stringify(files));
-            for(let i=0; i< files.length; i++) {
-                console.log("Index " + i);
-                console.log("File Name: " + files[i].name);
-                DriveHelper.deleteFile(files[i].name).then(() => {
-                    if(i === files.length - 1) window.location.reload();
-                });
+        DriveHelper.attemptFileRemovals().then((areFilesRemaining) => {
+            if(areFilesRemaining) {
+                DriveHelper.deleteAllFiles();
+            } else {
+                window.location.reload();
             }
-        }).catch(err => console.log("Error deleting all files: " + err));
+        }).catch((error) => {
+            console.log(error);
+            DriveHelper.deleteAllFiles();
+        })
+    }
+
+    static attemptFileRemovals() {
+        console.log("attempting removal");
+        return new Promise((resolve, reject) => {
+            DriveHelper.getFileList().then((files) => {
+                console.log("Files: " + JSON.stringify(files));
+                
+                if(JSON.stringify(files) === "[]") {
+                    resolve(false);
+                } else {
+                    for(let i=0; i < files.length; i++) {
+                        console.log("Index " + i);
+                        console.log("File Name: " + files[i].name);
+                        DriveHelper.deleteFile(files[i].name).then(() => {
+                            if(isNaN(files.length) || i === files.length - 1) {
+                                DriveHelper.getFileList().then((list) => {
+                                    resolve(list.length > 0);
+                                }).catch(error => reject(error));
+                            }
+                        }).catch(error => reject(error));
+                    }
+                }       
+            }).catch(error => reject("Error deleting all files: " + error));
+        });
     }
 
     /**
@@ -212,7 +232,7 @@ export default class DriveHelper {
                 // console.log(response); 
                 const files = response.result.files;
                 resolve(files.length);
-            }).catch(err => reject(err));
+            }).catch(error => reject(error));
         }); 
     }
 
@@ -232,7 +252,7 @@ export default class DriveHelper {
                 // console.log(response); 
                 const files = response.result.files;
                 resolve(files);
-            }).catch(err => reject(err));
+            }).catch(error => reject(error));
         }); 
     }
 }
